+++ src/app/components/molecules/edit-kwartet-card/edit-kwartet-card.component.scss
@@ -11,5 +11,5 @@
 }
 
 .border-overlay {
-  border-radius: 0.360rem
+  border-radius: 0.350rem
 }
