+++ src/app/components/pages/edit-set/edit-set.component.html
@@ -10,28 +10,28 @@
       <div
         class="flex flex-column align-items-center ml-3 mr-3">
         <p>Card 1</p>
-        <app-kwartet-card-preview [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card1"></app-kwartet-card-preview>
+        <app-edit-kwartet-card [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card1"></app-edit-kwartet-card>
         <p>{{currentEditedSet.card1.name}}</p>
       </div>
 
       <div
         class="flex flex-column align-items-center ml-3 mr-3">
         <p>Card 2</p>
-        <app-kwartet-card-preview [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card2"></app-kwartet-card-preview>
+        <app-edit-kwartet-card [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card2"></app-edit-kwartet-card>
         <p>{{currentEditedSet.card2.name}}</p>
       </div>
 
       <div
         class="flex flex-column align-items-center ml-3 mr-3">
         <p>Card 3</p>
-        <app-kwartet-card-preview [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card3"></app-kwartet-card-preview>
+        <app-edit-kwartet-card [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card3"></app-edit-kwartet-card>
         <p>{{currentEditedSet.card3.name}}</p>
       </div>
 
       <div
         class="flex flex-column align-items-center ml-3 mr-3">
         <p>Card 4</p>
-        <app-kwartet-card-preview [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card4"></app-kwartet-card-preview>
+        <app-edit-kwartet-card [kwartetSet]="currentEditedSet" [currentKwartetCard]="currentEditedSet.card4"></app-edit-kwartet-card>
         <p>{{currentEditedSet.card4.name}}</p>
       </div>
