+++ src/app/config/global-settings.ts
@@ -3,3 +3,6 @@ export const GAME_TITLE_CHARACTER_LIMIT: number = 20;
 
 // The character limit for Kwartet set titles
 export const SET_TITLE_CHARACTER_LIMIT: number = 16;
+
+// The character limit for Kwartet card names
+export const CARD_NAME_CHARACTER_LIMIT: number = 16;
